Move cleanup_device_registry() into util.py
